refactor: share auto reply helper fixtures
