+++ src/auto-reply/reply/test-helpers.ts
@@ -1,4 +1,5 @@
 import { vi } from "vitest";
+import type { FollowupRun } from "./queue.js";
 import type { TypingController } from "./typing.js";
 
 export function createMockTypingController(
@@ -14,5 +15,46 @@ export function createMockTypingController(
     markDispatchIdle: vi.fn(),
     cleanup: vi.fn(),
     ...overrides,
+  };
+}
+
+export function createMockFollowupRun(
+  overrides: Partial<Omit<FollowupRun, "run">> & { run?: Partial<FollowupRun["run"]> } = {},
+): FollowupRun {
+  const base: FollowupRun = {
+    prompt: "hello",
+    summaryLine: "hello",
+    enqueuedAt: Date.now(),
+    originatingTo: "channel:C1",
+    run: {
+      sessionId: "session",
+      sessionKey: "main",
+      messageProvider: "whatsapp",
+      agentAccountId: "primary",
+      sessionFile: "/tmp/session.jsonl",
+      workspaceDir: "/tmp",
+      config: {},
+      skillsSnapshot: {},
+      provider: "anthropic",
+      model: "claude",
+      thinkLevel: "low",
+      verboseLevel: "off",
+      elevatedLevel: "off",
+      bashElevated: {
+        enabled: false,
+        allowed: false,
+        defaultLevel: "off",
+      },
+      timeoutMs: 1_000,
+      blockReplyBreak: "message_end",
+    },
+  };
+  return {
+    ...base,
+    ...overrides,
+    run: {
+      ...base.run,
+      ...overrides.run,
+    },
   };
 }
